--- src/interfaces/gppc/test/gppc_test/expected/tablefunc.out
@@ -1,49 +1,0 @@
-CREATE FUNCTION tablefunc_describe(internal) RETURNS internal AS '$libdir/gppc_test' LANGUAGE c;
-CREATE FUNCTION tablefunc_project(anytable, int) RETURNS SETOF record AS '$libdir/gppc_test' LANGUAGE c WITH(describe=tablefunc_describe);
-CREATE FUNCTION describe_spi(internal) RETURNS internal AS '$libdir/gppc_test' LANGUAGE c;
-CREATE FUNCTION project_spi(anytable, text) RETURNS SETOF record AS '$libdir/gppc_test' LANGUAGE c WITH(describe=describe_spi);
-CREATE FUNCTION project_errorcallback(anytable, OUT int, OUT int) RETURNS SETOF record AS '$libdir/gppc_test' LANGUAGE c;
-SELECT * FROM tablefunc_project(TABLE(SELECT a, a / 10 FROM generate_series(1, 10)a SCATTER BY a), 2) ORDER BY 1;
-INFO:  attlen is 4
- ?column? 
-----------
-        0
-        0
-        0
-        0
-        0
-        0
-        0
-        0
-        0
-        1
-(10 rows)
-
-SELECT * FROM project_spi(TABLE(SELECT a::text FROM generate_series(1, 10)a SCATTER BY a), 'SELECT $$foo$$') ORDER BY 1;
-  foo  
--------
- foo1
- foo10
- foo2
- foo3
- foo4
- foo5
- foo6
- foo7
- foo8
- foo9
-(10 rows)
-
-SELECT * FROM project_errorcallback(TABLE(SELECT CASE WHEN a < 10 THEN a END, a FROM generate_series(1, 10)a SCATTER BY a));
-INFO:  message: Hello, world!  (seg2 slice2 172.17.0.4:25434 pid=19830)
-ERROR:  first attribute is NULL (gppc.c:1340)  (seg2 slice2 172.17.0.4:25434 pid=19830)
-SELECT * FROM project_errorcallback(TABLE(SELECT a, a FROM generate_series(1, 5)a SCATTER BY a));
- column1 | column2 
----------+---------
-       3 |       3
-       4 |       4
-       5 |       5
-       1 |       1
-       2 |       2
-(5 rows)
-
