--- src/interfaces/gppc/test/tabfunc_gppc_demo/init_file
@@ -1,9 +1,0 @@
--- start_matchignore
- m/NOTICE:  .*does not exist, skipping/
- m/NOTICE:  Table doesn\'t have \'DISTRIBUTED BY\' clause.*/
- m/HINT:  The \'DISTRIBUTED BY\' clause determines the distribution of data.*/
--- end_matchignore
--- start_matchsubs
-m/\(gppc\.c:[0-9]+\)/
-s/\(gppc\.c:[0-9]+\)//
--- end_matchsubs
